plant-disease-app-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
